Refactoring, moved the PrioritySorterJobColumn
 * everything in hudson.* is deprecated
